NestedNav link issues fixed. New banner component added. Other styling updates added.
