+++ src/components/shared/trending-banner/component-props.interface.ts
@@ -1,0 +1,10 @@
+export interface TrendingBannerComponentProps {
+  /**
+   * Banner heading
+   */
+  heading?: string;
+  /**
+   * Banner text
+   */
+  desc: string;
+}
