+++ src/components/trending/developers/component.tsx
@@ -2,9 +2,9 @@ import Box from "@primer/components/lib/Box";
 import DeveloperCardComponent from "../developer-card/component";
 import { DevelopersComponentProps } from "./component-props.interface";
 import Loading from "../../shared/loading/component";
-import TrendingToolbarComponent from "../../shared/trendingToolbar/component";
+import TrendingBannerComponent from "../../shared/trending-banner/component";
+import TrendingToolbarComponent from "../../shared/trending-toolbar/component";
 import useDevelopers from "../../../hooks/useDevelopers";
-
 /**
  * React Clone of Github Trending Developers https://github.com/developers
  * @typedef DevelopersComponentProps
@@ -13,20 +13,9 @@ export default function DevelopersComponent({ url }: DevelopersComponentProps) {
   const { status, data, error, isFetching } = useDevelopers();
 
   return (
-    <div>
-      <Box
-        bg="bg.tertiary"
-        borderColor="border.primary"
-        borderBottomWidth={1}
-        borderBottomStyle="solid"
-        p={6}
-        className="text-center"
-      >
-        <h1 className="h1">Trending</h1>
-        <p className="f4 color-text-secondary col-md-10 mx-auto">
-          These are the developers building the hot tools today.
-        </p>
-      </Box>
+    <>
+      {/* Banner */}
+      <TrendingBannerComponent desc="These are the developers building the hot tools today." />
 
       <section className="position-relative container-lg p-responsive pt-6">
         <Box
@@ -35,18 +24,11 @@ export default function DevelopersComponent({ url }: DevelopersComponentProps) {
           borderStyle="solid"
           borderRadius={6}
         >
-          <Box
-            bg="bg.tertiary"
-            borderColor="border.primary"
-            borderBottomWidth={1}
-            borderBottomStyle="solid"
-            p={3}
-            borderTopLeftRadius={6}
-            borderTopRightRadius={6}
-          >
-            <TrendingToolbarComponent type="developers" url={url} />
-          </Box>
-          <div>
+          {/* Toolbar */}
+          <TrendingToolbarComponent type="developers" url={url} />
+
+          {/* Developers Table */}
+          <>
             {status === "loading" ? (
               <Loading />
             ) : status === "error" ? (
@@ -68,9 +50,9 @@ export default function DevelopersComponent({ url }: DevelopersComponentProps) {
                 <div>{isFetching ? <Loading /> : " "}</div>
               </>
             )}
-          </div>
+          </>
         </Box>
       </section>
-    </div>
+    </>
   );
 }
