+++ src/components/trending/repository-card/component.tsx
@@ -5,7 +5,7 @@ import { RepositoryCardProps } from "./component-props.interface";
 export default function RepositoryCardComponent(props: RepositoryCardProps) {
   const { repo } = props;
   return (
-    <div>
+    <>
       <h1 className="h3 lh-condensed">
         <Link href={repo.url} target="_blank">
           <RepoIcon size={16} className="mr-1 color-text-secondary" />
@@ -13,6 +13,6 @@ export default function RepositoryCardComponent(props: RepositoryCardProps) {
           {repo.repositoryName}
         </Link>
       </h1>
-    </div>
+    </>
   );
 }
